Update null handling logic for projected dates (Recidiviz/recidiviz-data#40529)
## Description of the change

This PR updates the approach we take to handle null projected date
values in the sentencing v2 infrastructure. In Recidiviz/recidiviz-data#38538, I updated logic
to preserve NULLs in two distinct scenarios:
1. Within a sentence or sentence group, when an update is made that
nulls out a projected date, treat this null value as intentional and
preserve it over the previous non-null date.
2. When aggregating sentences or sentence groups to sentence inferred
groups, if more than one are being served simultaneously, preserve null
dates over non-null dates if there is one.

We have now determined that the second scenario is incorrect and that we
are nulling out values that we should not be when aggregating to the
inferred group level. This PR tweaks the logic in the second scenario
and keeps the first scenario as is.

This PR updates the logic to be consistent with how v1 sentences are
handled in supervision projected completion date spans and incarceration
projected completion date spans. When more than one sentence or sentence
group is being served simultaneously, non-null dates are prioritized
over non-null dates with one exception: if a sentence flagged as a life
sentence has a null projected completion date max value, we prioritize
the null value. In all other cases we preserve the non-null value.

The bigquery emulator tests for inferred groups are updated to test for
these specific scenarios.

The following
[notebook](https://colab.research.google.com/drive/1DVZOOtl2A8i_fOfSRkfBSBokSLSMWmTl?usp=sharing
) validates the output and confirms that there are no unexpected changes
to criteria or tasks as a result of this change.

We do see that the US_ID criteria that are built off of the
sessions_v2_all dataset do change (see screenshot below), but I believe
this is expected and addresses the issue that @samanthahuff surfaced in
Recidiviz/recidiviz-data#40484.

<img width="1455" alt="Screenshot 2025-04-04 at 11 12 40 AM"
src="https://github.com/user-attachments/assets/bf7a5a5c-df68-4c80-ab73-bcac09f48e2a"
/>

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#40484

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: de85010a9207e471a857854e8dac80e2e980649d