[US_AZ] Ingest charges from county (Recidiviz/recidiviz-data#40193)
## Description of the change

Probation is managed at the county level in AZ, and no probation
sentence information appears in the ADCRR state database. The only piece
of information we have about sentences that did not result in a
commitment to ADCRR custody lives in the `DOC_CLASS_ARREST_HISTORY`
table, which contains charges and offense dates for all arrests made
that did not later result in a commitment to ADCRR custody.

We need to ingest these charges so that we have a full set of
potentially exclusionary offenses on a person's record to gauge their
eligibility for certain opportunities.

Since we have only charge information and no additional sentence
information for these offenses, the ingest view in this PR creates
placeholder sentences with the necessary fields inferred from the charge
data. This will attach the charge to the correct person in
`state_charge_v2`, even though it will not attach those charges to a
meaningful sentence. These are all historical, so the serving period of
the sentence should never be relevant.

[sandbox
pipeline](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-03-31_15_42_34-14792751720568892712;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))&inv=1&invt=Abtilg)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e9b738ce3721f22a4802a1af5aa8c9dffd1293d8