[US_TN] Cleaning StaffSupervisorPeriod (Recidiviz/recidiviz-data#40586)
## Description of the change

Ingest pipeline failed with a supervisor period end date of
'1900-01-01'. We set the first known periods to have this date and one
externalID entry for a staff was lowercase so the view identified it as
a different person when evaluating their first supervisor period and had
the end date set to the start date.

The raw data has been fixed but I cleaned up the view a bit so it will
be less error prone since the raw data is created and updated manually,
and there is a tight feedback look with the tool, we do not want ingest
to be failing so easily and blocking data for other tools.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0381cde1189b9e224419043978055b2a9fc79302