[US_NC] Reset NC ingest logic (keeping raw data configs) (Recidiviz/recidiviz-data#40546)
## Description of the change

As we get closer to actually doing real ingest / product development
with NC, I think it's time to delete the legacy ingest views to:
a) avoid confusion
b) allow us to clean up some "reasonable date" exemptions I added for
NC.

This PR resets the NC ingest region mostly to factory settings, keeping
the configs for the legacy raw data round in case we want to reference
it later for some reason. I figure we can link this PR somewhere (maybe
on Recidiviz/recidiviz-data#34198?) so it's easy to reference this logic later if it's useful.

Once this PR is deployed, I will:
- [ ] Delete the `us_nc_ingest_view_results` / `us_nc_state` and
`us_nc_normalized_state` datasets in staging (they will get regenerated
as empty)
- [ ] Add a prod deploy checklist item to do the same

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#40488
Closes Recidiviz/recidiviz-data#19861

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d465e8275b85ebf03750d06a0c4cfdd7ae0e1b0c