[Insights] Modify Insights backend to source include_in_outcomes value from metric rows (Recidiviz/recidiviz-data#40495)
## Description of the change

Modifies the insights BE to query for `include_in_outcomes` metric rows
rather than view columns and removes obsolete view/schema remnants.

This PR is part 2 of 2 to start reading the officer entity's
`include_in_outcomes` value off of metric rows, rather than a
`include_in_outcomes` column. The original PR to do this was reverted
because we didn't separate the BE change from the view change - and the
BE was not backwards compatible. That PR has a very in depth description
of the reasoning behind this change: Recidiviz/recidiviz-data#40158.

This PR handles the querier changes which will be reading from the view
updates in the prior PR: Recidiviz/recidiviz-data#40494. This should not be merged until the
prior PR goes to prod. It switches up querier logic in a way that
depends on `include_in_outcomes` rows being present in the SOM view. It
also removes the now unnecessary `include_in_outcomes` columns and
schema fields.

Sandbox is at `gina3`, confirmed the old columns are removed, and we
still have a inclusion metric row for every officer.

Also diffed this against the reverted commit to be sure the changes
lined up!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#38094

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 42a1e886e6949c646e84308d14a25d91f2165efd