[Email reminders] Add menu for triggering email send (Recidiviz/recidiviz-data#40679)
## Description of the change

Adds a custom menu to the email reminders spreadsheet that will run the
functions that send line staff/supervisor emails, and makes those
functions private so that they can only be run from the menu. For now,
the menu items run the test email send functions--will replace those
with the real functions before merging.

I've been working in [TEST_ABC] if you want to see it in the Apps Script
UI.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#40428

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 330edf1a39a4f0f4dc37ebbd842fadac6d2287d7