[Testing] Randomize strings, names, and birthdays for fixture data (Recidiviz/recidiviz-data#40719)
## Description of the change

This PR creates new randomization capabilities for PII, which currently
comprises of names, external ID strings, other strings, and birthdate
columns.

I introduce a TODO to create a birthdate column type to tighten up how
we manage this special case, as it is the only datetime column we mark
as PII.

For the actual randomization, we set Faker's random seed based on the
value being passed in. This ensures that the random value is the same
across files and runs. (For example, we want `_DOC` values to be the
same across tests/fixture files if the actual value is the same, since
it represents the same value and is intended to be merged). Since the
actual value is never seen (I took precautions to avoid leaking info in
tests), I felt like this was an acceptable approach. Keeping it with the
consistent seed of 0 would lead to lots of problems (name clashing
and/or no reproducible randomized values, depending on the the column
and how many fixtures were pulled).

In short:
- Fully numeric strings are kept fully numeric, but random.
- Other strings are just random characters, but the same length as the
original value.
- Birthdates are moved a random amount within a year of the actual date
(to ensure they remain reasonable). However, this is preserved by the
datetime itself and consistent across the datetime format!

I included a little script that prints out PII column information across
all raw data configs, which accelerated this task (not sure if anyone
else will need it... but leaving it here to expand over time as needed).
This did reveal one INTEGER column in OZ that should probably be an
external ID type. But that can be handled when I test fixture generation
in OZ.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| **Type: Feature** | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#40159

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8ac4daa833a4cc219479ad3c433f98ba932a2105