[US_TN] Adding filter to only have IP results if someone exists in OffenderMovement table (Recidiviz/recidiviz-data#40645)
## Description of the change

If TN deletes someone from their data they will have no record in
OffenderMovement table. This PR adds a filter so we will also not
surface IP results for those cases.

This view brings the results from 8026292 to 8026244, which is a
difference of 48 periods. The case that surfaces this issue had 2
periods, and we expect there are other instances of this as well. We
feel okay about this change since we don't expect someone to ever be
incarcerated in TN without any record in OffenderMovement.

Successful ingest pipeline
[here](https://console.cloud.google.com/dataflow/jobs?project=recidiviz-123&authuser=0&invt=AbuJsg).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#7800 and Recidiviz/recidiviz-data#40639

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 844ae04db171a9edc5997da21af36308bf6af6bc