don't truncate marked ineligible spans at reset dates (Recidiviz/recidiviz-data#40576)
## Description of the change

If a user is snoozed while eligible, then becomes ineligible internally,
then becomes eligible again, all before the snooze expires, we would
show them as marked ineligible in the product. This updates the funnel
view to match that logic.

Loaded to sandbox at `dana_funnel`, the data in staging is not super
indicative of real marked ineligibles but a spot check looks right to
me, and I didn't feel like spending the time trying to figure out how to
make a prod sandbox work with `auto` mode when loading views.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#40435

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f1f5400a950dbcd098dceb844deafbbe17e05c30