[US_PA][Worfklows] Update form query to account for unreported dispositions (Recidiviz/recidiviz-data#40313)
## Description of the change

This PR adjusts the logic to auto-fill the [402A
form](https://drive.google.com/file/d/1BEshwzVh1m5MRBfAQJivZ9jFfyyW2Qxj/view?usp=sharing)
for admin supervision to account for unreported dispositions (which we
recently ingested).

Specifically, it creates a new variable called
`form_information_drug_unreported_disposition` that will auto-fill the
second checkbox in section 1 of the 402a form. It also renames
`form_information_drug_charge` to `form_information_drug_conviction` for
auto-filling the first checkbox in section 1 of the 402a form and adds
logic to make sure that we only include convictions. Finally, it fixes a
typo where I misspelled statue as statute 😅 and makes sure that all the
`form_information_statute` variables that populate section 2 of the form
are checked for unreported dispositions OR convictions for those
charges.

EDIT: per polaris request I also kept in all the old variable names and
made a TODO to remove them later, so we don't break anything on the
front end.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#39074

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 509edc1f0ff5386d169ea07adbf89f7fe41f9eac