Add more incarceration starts and absconsion metrics to insights impact explore (Recidiviz/recidiviz-data#40579)
## Description of the change

For the supervisor homepage dashboard, we want to be able to show the
relevant incarceration starts and absconsions metrics for each state
where the outcomes module is live.

This PR adds all of the metrics in the [outliers
configs](https://github.com/Recidiviz/recidiviz-data/blob/main/recidiviz/outliers/outliers_configs.py)
to the insights impact explore.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Advances https://github.com/Recidiviz/looker/issues/696

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@recidiviz.org>
GitOrigin-RevId: 0570437dcd9a2c808fab6db826dd1faf4e5ea83a