[US_AR] Use INNER join in both sentence status and sentence length queries (Recidiviz/recidiviz-data#40759)
## Description of the change

Recidiviz/recidiviz-data#40310 changed the sentence status view to use an INNER join on
SENTENCECOMPONENT to fix errors that were being caused by the LEFT join.
The same change was not made to the sentence lengths view, so sentences
lacking SENTENCECOMPONENT data were still being ingested based on that
view.

This change moves the INNER join to the sentence lengths view query; it
also reverts the change in Recidiviz/recidiviz-data#40310, since sentence statuses pull from the
sentence lengths fragment, so the change to the sentence lengths
fragment will fix the issue for both views.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5e46c4d54e675cdad1a225a07f0388ec626f0962